Separate out overflow-group into OverflowGroupMixin and overflow-group component
